更改调试数据发送，增加读取驱动器温度，读取单片机复位标志，需要改进优化

+++ 主控1.0.0.161231_α/Action_User/GET_SET.h
@@ -35,8 +35,8 @@ float GetAngle(void);
 void UpdateVel(void);
 float GetVel(void);
 
-void SetMotorVel(float *value);
-float GetMotorVel(int motorNum);
+void SetMotorVel(uint8_t *value);
+uint8_t GetMotorVel(int motorNum);
 float GetMotorVelRate(void);
 float GetMotorVelAng(void);
